+++ app/src/main/java/me/proteus/myeye/MainActivity.kt
@@ -19,13 +19,13 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import me.proteus.myeye.ui.DistanceMeasurement
-import me.proteus.myeye.ui.MapScreen
-import me.proteus.myeye.ui.PlaceDetailsScreen
-import me.proteus.myeye.ui.ResultBrowserScreen
-import me.proteus.myeye.ui.SettingsScreen
-import me.proteus.myeye.ui.TestResultScreen
-import me.proteus.myeye.ui.VisionTestScreen
+import me.proteus.myeye.ui.screens.DistanceMeasurement
+import me.proteus.myeye.ui.screens.MapScreen
+import me.proteus.myeye.ui.screens.PlaceDetailsScreen
+import me.proteus.myeye.ui.screens.ResultBrowserScreen
+import me.proteus.myeye.ui.screens.SettingsScreen
+import me.proteus.myeye.ui.screens.TestResultScreen
+import me.proteus.myeye.ui.screens.VisionTestScreen
 import me.proteus.myeye.ui.components.TestSelector
 import me.proteus.myeye.ui.theme.MyEyeTheme
 import me.proteus.myeye.util.LanguageUtils
