+++ app/src/main/java/me/proteus/myeye/ui/activities/DistanceTrackerActivity.kt
@@ -1,4 +1,4 @@
-package me.proteus.myeye.ui
+package me.proteus.myeye.ui.activities
 
 import android.hardware.Sensor
 import android.hardware.SensorEvent
