+++ app/src/main/java/me/proteus/myeye/ui/screens/ResultBrowserScreen.kt
@@ -1,4 +1,4 @@
-package me.proteus.myeye.ui
+package me.proteus.myeye.ui.screens
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
