+++ app/src/main/java/me/proteus/myeye/ui/screens/TestResultScreen.kt
@@ -1,4 +1,4 @@
-package me.proteus.myeye.ui
+package me.proteus.myeye.ui.screens
 
 import android.annotation.SuppressLint
 import androidx.biometric.BiometricPrompt
@@ -26,6 +26,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import me.proteus.myeye.TestResult
 import me.proteus.myeye.io.ResultDataSaver
+import me.proteus.myeye.ui.AuthorizationViewModel
 import me.proteus.myeye.ui.components.VisionTestIcon
 import me.proteus.myeye.ui.theme.MyEyeTheme
